feat: implement account management functions; add add, update, delete, and retrieval operations for accounts, and integrate with Supabase
